requirements.txt
anyio==3.6.1
asgiref==3.5.2
autopep8==1.6.0
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
MarkupSafe==2.1.1
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pydotenv==0.0.7
PyJWT==2.4.0
pymongo==4.1.1
PyMySQL==1.0.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
toml==0.10.2
typing-extensions==4.2.0
uvicorn==0.17.6
